Add all the user claims to meta endpoint.
